# backend/src/main/java/com/dsw02/empleados/model/DepartamentoResponse.java
package com.dsw02.empleados.model;

public class DepartamentoResponse {

    private String clave;
    private String nombre;

    public String getClave() {
        return clave;
    }

    public void setClave(String clave) {
        this.clave = clave;
    }

    public String getNombre() {
        return nombre;
    }

    public void setNombre(String nombre) {
        this.nombre = nombre;
    }
}