# backend/src/main/java/com/dsw02/empleados/model/DepartamentoUpdateRequest.java
package com.dsw02.empleados.model;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

@JsonIgnoreProperties(ignoreUnknown = false)
public class DepartamentoUpdateRequest {

    @NotBlank
    @Size(max = 100)
    private String nombre;

    public String getNombre() {
        return nombre;
    }

    public void setNombre(String nombre) {
        this.nombre = nombre;
    }
}