# specs/001-departamentos-empleados-crud/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Empleados y Departamentos API v1
  version: 1.0.0
  description: CRUD versionado de empleados y departamentos con Basic Auth por roles
servers:
  - url: http://localhost:8080
security:
  - basicAuth: []
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    DepartamentoCreateRequest:
      type: object
      required: [nombre]
      properties:
        nombre:
          type: string
          maxLength: 100
    DepartamentoUpdateRequest:
      type: object
      required: [nombre]
      properties:
        nombre:
          type: string
          maxLength: 100
    DepartamentoResponse:
      type: object
      properties:
        clave:
          type: string
        nombre:
          type: string
    EmpleadoCreateRequest:
      type: object
      required: [nombre, direccion, telefono, email, password, role]
      properties:
        nombre:
          type: string
        direccion:
          type: string
        telefono:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [ADMIN, USER]
        departamentoClave:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
paths:
  /api/v1/departamentos:
    get:
      summary: Listar departamentos paginados
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - in: query
          name: sort
          schema:
            type: string
            default: clave,asc
      responses:
        '200':
          description: Lista paginada
        '400':
          description: Parametros invalidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Crear departamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartamentoCreateRequest'
      responses:
        '201':
          description: Departamento creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartamentoResponse'
        '403':
          description: Forbidden
  /api/v1/departamentos/{clave}:
    get:
      summary: Obtener departamento por clave
      parameters:
        - in: path
          name: clave
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Departamento encontrado
        '404':
          description: No encontrado
    put:
      summary: Actualizar departamento
      parameters:
        - in: path
          name: clave
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Actualizado
        '403':
          description: Forbidden
    delete:
      summary: Eliminar departamento
      parameters:
        - in: path
          name: clave
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Eliminado
        '409':
          description: Tiene empleados asociados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'